+++ herostratus/tests/check.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
 use herostratus::git::clone::find_local_repository;
-use herostratus_tests::cmd::{CommandExt, herostratus};
+use herostratus_tests::cmd::{CommandExt, exclude_all_rules_except, herostratus};
 use predicates::prelude::*;
 use predicates::str;
 
@@ -26,8 +26,39 @@ fn search_current_repo_for_branch_that_does_not_exist() {
 }
 
 #[test]
+fn search_depth() {
+    let config = exclude_all_rules_except("H1-fixup");
+    let (mut cmd, _temp) = herostratus(None, Some(config.clone()));
+    // The fixup branch's HEAD is not a fixup commit, but its parent is.
+    cmd.arg("check")
+        .arg(".")
+        .arg("origin/test/fixup")
+        .arg("--depth=1");
+    let output = cmd.captured_output();
+    let stderr = String::from_utf8_lossy(&output.stderr); // herostratus logs to stderr
+    assert!(output.status.success());
+
+    let assertion = str::contains("processing 1 commits");
+    assert!(assertion.eval(&stderr), "Found != 1 commits");
+
+    let (mut cmd, _temp) = herostratus(None, Some(config));
+    cmd.arg("check")
+        .arg(".")
+        .arg("origin/test/fixup")
+        .arg("--depth=2");
+
+    let output = cmd.captured_output();
+    let stderr = String::from_utf8_lossy(&output.stderr);
+    assert!(output.status.success());
+
+    let assertion = str::contains("processing 2 commits");
+    assert!(assertion.eval(&stderr), "Found != 2 commits");
+}
+
+#[test]
 fn search_current_repo_for_fixup_commits() {
-    let (mut cmd, _temp) = herostratus(None, None);
+    let config = exclude_all_rules_except("H1-fixup");
+    let (mut cmd, _temp) = herostratus(None, Some(config));
     cmd.arg("check").arg(".").arg("origin/test/fixup");
 
     let output = cmd.captured_output();
@@ -45,6 +76,7 @@ fn search_current_repo_for_fixup_commits() {
 ///
 /// The application should not crash nor exit with an error on any branch.
 #[test]
+#[cfg_attr(not(feature = "ci"), ignore = "Slow test, only run in CI")]
 fn smoke_test_on_all_own_branches() {
     use std::os::unix::ffi::OsStringExt;